pyproject.toml
[project]
name = "events-planner-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastapi-mcp>=0.3.7",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.0",
    "python-dotenv>=1.1.1",
    "resend>=2.11.0",
    "uvicorn>=0.35.0",
]
